requirements.txt
streamlit
google-generativeai
python-dotenv
crewai
crewai_tools
chromadb
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
langchain-google-genai
pysqlite3-binary
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq
langchain_core
fastapi[standard]
uvicorn
langserve[all]
sse_starlette
nbconvert
